Actually make the PS3 stuff work
Haven't tested on a real PS3, but have tested on the hack of RPCS3 that fixes the game's quirks and it seemed to be working?

Model Randomisation is less feature rich in the PS3 version due to the lack of SoXMDL support in Marathon.

Do also want to try and add Custom support to the PS3 version as well, but I'll need a way to encode AT3 audio (the voice packs will all need updating to have AT3 sound files alongside the XMAs).